put css

+++ src/components/BoardList/list.tsx
@@ -1,10 +1,10 @@
 import React, { ReactElement, MouseEvent } from "react";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { makeStyles } from "@material-ui/core/styles";
 import CloseIcon from "@material-ui/icons/Close";
+import { IIsOnState, IrootState, IBoard } from "../../store/_type";
 import { reqDeleteBoard } from "../../store/_action/board";
 import { reqUpdateIsOn } from "../../store/_action/isOn";
-import { IBoard } from "../../store/_type";
 
 interface ListProps {
     board: IBoard;
@@ -12,15 +12,28 @@ interface ListProps {
 
 export default function List({ board }: ListProps): ReactElement {
     const dispatch = useDispatch();
+    const isOn: IIsOnState = useSelector((state: IrootState) => state.isOn);
     const useStyles = makeStyles({
-        listItem: {
-            width: "100%",
-            height: "2.5em",
+        listWrapper: {
+            display: "flex",
+            justifyContent: "space-between",
+            alignItems: "center",
             color: "#ffffff",
             fontWeight: 500,
             fontSize: "1.2em",
-            padding: "10px 5px",
+            width: "100%",
+            height: "2.5em",
+            padding: ".5em",
             cursor: "pointer",
+            "&:hover": {
+                backgroundColor: "#334280",
+            },
+        },
+        selectedList: {
+            backgroundColor: "#334280",
+        },
+        deleteBtn: {
+            color: "#7e889a",
         },
     });
     const handleBoardOn = (e: MouseEvent) => {
@@ -30,20 +43,27 @@ export default function List({ board }: ListProps): ReactElement {
     const handleDeleteBoard = (e: MouseEvent) => {
         e.stopPropagation();
         e.preventDefault();
-        dispatch(reqDeleteBoard(board.timestamp));
+        if (window.confirm("정말 삭제하시겠습니까?")) {
+            dispatch(reqDeleteBoard(board.timestamp));
+        }
     };
     const classes = useStyles();
     return (
-        <>
-            <li
-                id={`${board.timestamp}`}
-                className={classes.listItem}
-                key={`boardIndex-${board.index}`}
-                onClick={handleBoardOn}
-            >
-                {board.name}
-                <CloseIcon onClick={handleDeleteBoard} />
-            </li>
-        </>
+        <li
+            id={`${board.timestamp}`}
+            className={
+                isOn.boardId === board.timestamp
+                    ? `${classes.listWrapper} ${classes.selectedList}`
+                    : `${classes.listWrapper}`
+            }
+            key={`boardIndex-${board.index}`}
+            onClick={handleBoardOn}
+        >
+            {board.name}
+            <CloseIcon
+                className={classes.deleteBtn}
+                onClick={handleDeleteBoard}
+            />
+        </li>
     );
 }
